+++ backend/learn-api/src/main/java/com/prosper/learn/api/client/MessageClient.java
@@ -12,20 +12,20 @@ public interface MessageClient {
 
     //@PostMapping("/message")
     Response create(@RequestParam("content") String content,
-                    @RequestParam("senderId") int senderId,
-                    @RequestParam("receiverId") int receiverId);
+                    @RequestParam("senderId") Long senderId,
+                    @RequestParam("receiverId") Long receiverId);
 
     //@GetMapping("/message/{id}")
-    Response<MessageDTO> get(@PathVariable int id);
+    Response<MessageDTO> get(@PathVariable Long id);
 
     @GetMapping("/message/system")
     Response<Map<String, Object>> getSystemList(
-            @RequestParam int type, @RequestParam("userId") int receiverId, @RequestParam int lastId);
+            @RequestParam int type, @RequestParam("userId") Long receiverId, @RequestParam Long lastId);
 
     @GetMapping("/message/course-apply")
-    Response<Map<String, Object>> getCourseApplyList(@RequestParam("userId") int senderId, @RequestParam int lastId);
+    Response<Map<String, Object>> getCourseApplyList(@RequestParam("userId") Long senderId, @RequestParam Long lastId);
 
     @PostMapping("/message/invite")
-    Response Invite(@RequestParam int userId, @RequestParam int nodeId);
+    Response Invite(@RequestParam Long userId, @RequestParam Long nodeId);
 
 }
